+++ deploy.sh
@@ -139,6 +139,7 @@ JS_FILES=(
   i18n_ro.js
   iec-symbols.js
   mounting-svg.js
+  panel-svg.js
   pt-doc.js
   pt-text-ro.js
   pt-text-en.js